% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute}
\alias{compute}
\title{Computing TPR and PPV}
\usage{
compute(ts_de, in_de)
}
\arguments{
\item{ts_de}{a list of new truth sets attained from function "truthset_de".}

\item{in_de}{a list of results of subgroup DESeq2 analyses.}
}
\value{
a list of vectors which contain corresponding TPR values and PPV values.
}
\description{
This function computes TPR(true positive rate) and PPV(positive predictive value)
according to the results we obtained from DESeq2 analysis.
}
\examples{
compute(ts_de, in_de)
}
