% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasplit_de.R
\name{datasplit_de}
\alias{datasplit_de}
\title{Save splitted data}
\usage{
datasplit_de()
}
\description{
This function loads data sets from package "mydata", splits them into 
separated text files and saves in "/data" subdirectory for DESeq2 analysis.
}
\examples{
datasplit_de()
}
