% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truthset_de.R
\name{truthset_de}
\alias{truthset_de}
\title{Truth Set Redefinition}
\usage{
truthset_de(N, n, sam, datatype)
}
\arguments{
\item{N}{integer. The number of iterations for resampling the original
truth set.}

\item{n}{integer. The number of samples in the new truth set.}

\item{sam}{matrix obtained from first-step resampling.}

\item{datatype}{character. "luad", "brca" or "sarc".}
}
\value{
a list which is composed of N new truth sets for N iterations, 
        respectively.
}
\description{
The function is used to obtain a new truth set for different 
             data types via a combination of certain variables.
}
\details{
This function imports a matrix attained from the first-step resampling,
compares the DESeq results with that of the original truth set and exports
a group of new truth sets(gene lists).
Note that N should be equal to the number of rows of matrix sam,
n should be equal to the 2 times the number of columns of matrix sam("luad",
"brca") and 1 time the number columns of matrix sam("sarc").
}
\examples{
truthset_de(100, 170, sam, datatype="brca")
}
\keyword{DESeq2}
