\name{varianceFitDiagnostics}
\Rdversion{1.1}
\alias{varianceFitDiagnostics}

\title{
REMOVED
}
\description{
This function has been removed. Please see the vignette for our newer suggestions on
how to check fit quality.
}
\usage{
varianceFitDiagnostics( ... )
}
\arguments{
  \item{...}{
      dummy argument}
}

