\name{CountDataSet-class}
\Rdversion{1.1}
\docType{class}
\alias{CountDataSet-class}

\title{Class "CountDataSet" -- a container for count data from HTS experiments }
\description{	This is the main class for the present package. }
\section{Objects from the Class}{
Objects should be created with calls to \code{\link{newCountDataSet}} (q.v.).
}
\section{Extends}{
Class \code{eSet} (package 'Biobase'), directly.
Class \code{VersionedBiobase} (package 'Biobase'), by class "eSet", distance 2.
Class \code{Versioned} (package 'Biobase'), by class "eSet", distance 3.
}
\note{ For a description of the class and its slots, see the appendix of the vignette. }

\examples{
# See the vignette
}

