\name{varAdjFactors}
\Rdversion{1.1}
\alias{varAdjFactors}
\alias{varAdjFactors<-}
\title{
    Defunct function. Do not use.
}
\description{
   For a few version (which were only available in the deval branch),
   the possibility was given to scale the variance estimates if desired.
   As this turned out to cause more design issues than it solved, the
   fucntionality was removed again.

}
\author{
   Simon Anders, sanders@fs.tum.de
}

